+++ phono3py/cui/load.py
@@ -104,7 +104,7 @@ def load(
     is_symmetry: bool = True,
     symmetrize_fc: bool = True,
     is_mesh_symmetry: bool = True,
-    is_compact_fc: bool = False,
+    is_compact_fc: bool = True,
     use_pypolymlp: bool = False,
     mlp_params: dict | None = None,
     use_grg: bool = False,
@@ -251,12 +251,13 @@ def load(
         True.
     is_compact_fc : bool, optional
         fc3 are created in the array whose shape is
-            True: (primitive, supercell, supercell, 3, 3, 3) False: (supercell,
-            supercell, supercell, 3, 3, 3)
+            True: (primitive, supercell, supercell, 3, 3, 3)
+            False: (supercell, supercell, supercell, 3, 3, 3)
         and for fc2
-            True: (primitive, supercell, 3, 3) False: (supercell, supercell, 3, 3)
+            True: (primitive, supercell, 3, 3)
+            False: (supercell, supercell, 3, 3)
         where 'supercell' and 'primitive' indicate number of atoms in these
-        cells. Default is False.
+        cells. Default is True.
     use_pypolymlp : bool, optional
         Use pypolymlp for generating force constants. Default is False.
     mlp_params : dict, optional
@@ -404,6 +405,7 @@ def load(
             fc_calculator_options=fc_calculator_options,
             symmetrize_fc=symmetrize_fc,
             is_compact_fc=is_compact_fc,
+            use_symfc_projector=True,
         )
 
     if log_level and ph3py.fc3 is not None:
@@ -439,7 +441,7 @@ def compute_force_constants_from_datasets(
     cutoff_pair_distance: float | None = None,
     symmetrize_fc: bool = True,
     is_compact_fc: bool = True,
-    load_phono3py_yaml: bool = False,
+    use_symfc_projector: bool = False,
 ) -> None:
     """Compute force constants from datasets.
 
@@ -466,7 +468,7 @@ def compute_force_constants_from_datasets(
             is_compact_fc=is_compact_fc,
             fc_calculator=fc3_calculator,
             fc_calculator_options=fc3_calc_opts,
-            use_symfc_projector=load_phono3py_yaml,
+            use_symfc_projector=use_symfc_projector,
         )
 
     if ph3py.fc2 is None or fc3_calculator != fc2_calculator:
@@ -481,7 +483,7 @@ def compute_force_constants_from_datasets(
                 is_compact_fc=is_compact_fc,
                 fc_calculator=fc2_calculator,
                 fc_calculator_options=fc2_calc_opts,
-                use_symfc_projector=load_phono3py_yaml,
+                use_symfc_projector=use_symfc_projector,
             )