+++ phono3py/phonon3/real_to_reciprocal.py
@@ -1,44 +1,12 @@
 """Transform fc3 in real space to reciprocal space."""
 
-# Copyright (C) 2020 Atsushi Togo
-# All rights reserved.
-#
-# This file is part of phono3py.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions
-# are met:
-#
-# * Redistributions of source code must retain the above copyright
-#   notice, this list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright
-#   notice, this list of conditions and the following disclaimer in
-#   the documentation and/or other materials provided with the
-#   distribution.
-#
-# * Neither the name of the phonopy project nor the names of its
-#   contributors may be used to endorse or promote products derived
-#   from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
 from __future__ import annotations
 
 import numpy as np
 from numpy.typing import NDArray
 from phonopy.structure.cells import Primitive
+
+from phono3py.phonon3.fc3 import compact_fc3_to_full_fc3
 
 
 def run_real_to_reciprocal_rust(
@@ -110,7 +78,11 @@ class RealToReciprocal:
         ``byte``) is required when ``make_r0_average`` is True.
 
         """
-        self._fc3 = fc3
+        # Only full fc3 is supported in this python implementation.
+        if fc3.shape[0] != fc3.shape[1]:
+            self._fc3 = compact_fc3_to_full_fc3(primitive, fc3)
+        else:
+            self._fc3 = fc3
         self._primitive = primitive
         self._mesh = mesh
         self._symprec = symprec